Update inference code
