+++ src/inference_utils.py
@@ -1,6 +1,11 @@
+import shlex
+import subprocess
+import sys
+
 import torch
 from pytorch_pretrained_bert import BertTokenizer, \
     BertForSequenceClassification, BertForNextSentencePrediction, BertForMaskedLM
+
 
 def load_pretrained_model_tokenizer(base_model=None, base_tokenizer=None):
     # Load pre-trained model (weights)
@@ -42,3 +47,21 @@ def print_scores(scores, mode="test"):
     for sn, score in zip(scores[0], scores[1]):
         print("{}: {}".format(sn, score), end=" ")
     print("")
+
+
+def evaluate(trec_eval_path, predictions_file, qrels_file):
+    cmd = trec_eval_path + " {judgement} {output} -m map -m recip_rank -m P.30".format(
+        judgement=qrels_file, output=predictions_file)
+    pargs = shlex.split(cmd)
+    print("running {}".format(cmd))
+    p = subprocess.Popen(pargs, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    pout, perr = p.communicate()
+
+    if sys.version_info[0] < 3:
+        lines = pout.split(b'\n')
+    else:
+        lines = pout.split(b'\n')
+    map = float(lines[0].strip().split()[-1])
+    mrr = float(lines[1].strip().split()[-1])
+    p30 = float(lines[2].strip().split()[-1])
+    return map, mrr, p30
